Corrected the main to prevent constant semaphore acquiring/releasing that was freezing the message queue. Refactoring, adding some comments, minor improvements.

+++ Src/Devices/Actuators/FakeActuator.cpp
@@ -7,14 +7,14 @@ FakeActuator::FakeActuator() : digitalOut(UPinUtils::digitalInOut[0])
     m_actuatorName = "";
 }
 
-FakeActuator::FakeActuator(uint8_t a_id, uint8_t a_pin, char* a_actuatorName) : digitalOut(UPinUtils::digitalInOut[a_pin])
+FakeActuator::FakeActuator(int a_id, int a_pin, char* a_actuatorName) : digitalOut(UPinUtils::digitalInOut[a_pin])
 {
     m_actuatorId = a_id;
     m_actuatorName = a_actuatorName;
 }
 
 // Save the current state of the sensor.
-void FakeActuator::SetValue(uint8_t a_value)
+void FakeActuator::SetValue(int a_value)
 {
     digitalOut = a_value;
 }
